+++ src/draw.js
@@ -1,37 +1,40 @@
-/* global importPackage Packages player context argv */
 import { DOMParser, DOMImplementation } from 'xmldom' /* eslint-disable-line no-unused-vars */
 import toGeoJSON from 'togeojson'
 
+import { draw as usage } from './modules/usage'
 import decode from './modules/decodePolygon'
-import { draw, findGround, naturalBlock, oneBlockAbove, setBlock, printBlocks } from './modules/drawLines'
+import { draw, findGround, ignoreBuildings, setOffset, setWall, printBlocks } from './modules/drawLines'
 import { ignoredBlocks, allowedBlocks } from './modules/blocks'
 import { readFile } from './modules/readFile'
+import { transformIDs } from './modules/utils'
 
 importPackage(Packages.com.sk89q.worldedit)
 importPackage(Packages.com.sk89q.worldedit.math)
 importPackage(Packages.com.sk89q.worldedit.blocks)
 
-const usage = `/cs draw <file> <block> [flags]
- • §o/cs draw rails1 iron_block
- • §o/cs draw file3 stone u
-Flags:
- • §lu§r§c Draws a block above`
+context.checkArgs(1, 3, usage)
 
-context.checkArgs(2, 3, usage)
+const options = {
+  block: argv[2] || 'gold_block',
+  offset: 0,
+  height: 1,
+  onGround: true,
+  ignoreBuildings: true,
+  ignoreVegetation: true,
+  ignoredBlocks,
+  allowedBlocks,
+  ...JSON.parse(argv[3] || '{}')
+}
 
-const blocks = context.remember()
-
-const [block, flags] = argv.slice(2)
-
-const up = flags && ('' + flags).includes('u')
-const options = { block, up }
+transformIDs(options, 'ignoredBlocks')
+transformIDs(options, 'allowedBlocks')
 
 player.print('§7Please wait...')
 
-process(argv[1])
+process(options)
 
-function process (filename) {
-  const file = context.getSafeOpenFile('drawings', filename, 'geojson', ['json', 'geojson', 'kml'])
+function process (options) {
+  const file = context.getSafeOpenFile('drawings', argv[1], 'geojson', ['json', 'geojson', 'kml'])
   const data = readFile(file)
 
   if (!file.exists()) {
@@ -49,21 +52,22 @@ function process (filename) {
     drawing = toGeoJSON.kml(dom)
     player.print('§7Imported KML...')
   }
-  drawRaw(drawing)
+  drawRaw(drawing, options)
 }
 
-function drawRaw (data) {
+function drawRaw (data, options) {
   const lines = decode(data)
-  const findGround_ = findGround(ignoredBlocks, blocks)
-  const naturalBlock_ = naturalBlock(allowedBlocks, blocks)
-  const oneBlockAbove_ = oneBlockAbove(options)
-  const setBlock_ = setBlock(blocks, context, block)
+  const findGround_ = findGround(options)
+  const ignoreBuildings_ = ignoreBuildings(options)
+  const setOffset_ = setOffset(options)
+  const setWall_ = setWall(options)
   draw(lines, (pos) => {
     pos = findGround_(pos)
-    if (naturalBlock_(pos)) {
-      pos = oneBlockAbove_(pos)
-      setBlock_(pos)
+    if (ignoreBuildings_(pos)) {
+      pos = setOffset_(pos)
+      setWall_(pos)
     }
+    player.print(pos)
   })
   printBlocks()
 }