+++ src/gis_jp.js
@@ -1,5 +1,7 @@
+import { gis_jp as usage } from './modules/usage'
 import { elevation } from './modules/elevation'
 import { ignoredBlocks } from './modules/blocks'
+import { transformIDs } from './modules/utils'
 
 importClass(Packages.com.sk89q.worldedit.Vector)
 
@@ -11,11 +13,6 @@ importClass(Packages.java.lang.Thread)
 importClass(Packages.java.nio.charset.StandardCharsets)
 importClass(Packages.org.apache.commons.io.IOUtils)
 
-const usage = `[flags]
-Flags:
- • §lw§r§c Keeps water
- • §ls§r§c Removes smoothing step`
-
 context.checkArgs(0, 1, usage)
 
 const options = {
@@ -25,8 +22,9 @@ const options = {
   ...JSON.parse(argv[1] || '{}')
 }
 
-options.ignoredBlocks = options.ignoredBlocks.map((id) => context.getBlock(id).id)
+transformIDs(options, 'ignoredBlocks')
 
 elevation(options,
-  (lons, lats) => `http://wxs.ign.fr/${process.env.IGN_API_KEY}/alti/rest/elevation.json?lon=${lons}&lat=${lats}&zonly=true`,
-  (data) => data.elevations)
+  (lons, lats) => `https://cyberjapandata2.gsi.go.jp/general/dem/scripts/getelevation.php?lon=${lons}&lat=${lats}&outtype=JSON`,
+  (data) => [data.elevation],
+  1)