+++ src/modules/OSMcommand.js
@@ -1,7 +1,7 @@
 import overpass from './overpass'
 import getProjection from './getProjection'
 
-export function request (radius, center, getQuery, draw) {
+export function request (radius, center, getQuery, callback) {
   player.print('§7Please wait...')
   const points = transformPoints(getPoints(radius, center))
   const s = findS(points)
@@ -10,14 +10,8 @@ export function request (radius, center, getQuery, draw) {
 
   overpass(query, (err, data) => {
     if (err) throw err
-    draw(data)
+    callback(data)
   })
-}
-
-export function getRadius (region) {
-  const x = Math.abs(region.pos1.x - region.pos2.x)
-  const z = Math.abs(region.pos1.z - region.pos2.z)
-  return Math.sqrt(x * x + z * z) / 2
 }
 
 function getPoints (radius, center) {