+++ src/modules/blocks.js
@@ -1,5 +1,3 @@
-/* global context */
-
 export const ignoredBlocks = [
   'air',
   'tallgrass',
@@ -24,13 +22,14 @@ export const ignoredBlocks = [
 ]
 
 export const allowedBlocks = [
-  context.getBlock('grass').id,
-  context.getBlock('dirt').id,
-  context.getBlock('stone').id,
-  context.getBlock('sand').id,
-  context.getBlock('grass_path').id,
-  context.getBlock('concrete').id,
-  context.getBlock('gravel').id,
-  context.getBlock('water').id,
-  context.getBlock('lava').id
+  'air',
+  'grass',
+  'dirt',
+  'stone',
+  'sand',
+  'grass_path',
+  'concrete',
+  'gravel',
+  'water',
+  'lava'
 ]