+++ src/modules/colors.js
@@ -1,0 +1,33 @@
+export const black = '§0'
+export const darkBlue = '§1'
+export const darkGreen = '§2'
+export const darkAqua = '§3'
+export const darkRed = '§4'
+export const darkPurple = '§5'
+export const gold = '§6'
+export const gray = '§7'
+export const darkGray = '§8'
+export const blue = '§9'
+export const green = '§a'
+export const aqua = '§b'
+export const red = '§c'
+export const lightPurple = '§d'
+export const yellow = '§e'
+export const white = '§f'
+
+export const obscufated = '§k'
+export const bold = '§l'
+export const strikethrough = '§m'
+export const underline = '§n'
+export const italic = '§o'
+export const reset = '§r'
+
+export const command = gold
+export const required = green
+export const optional = aqua
+export const example = `${gray}${italic}`
+
+export const object = white
+export const number = red
+export const boolean = lightPurple
+export const string = yellow